+++ src/test/java/ru/netology/stats/javamvn7/services/StatsServiceTest.java
@@ -12,7 +12,7 @@ public class StatsServiceTest {
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         int expected = 9;
-        int actual = service.minSales(sales);
+        long actual = service.minSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -24,7 +24,7 @@ public class StatsServiceTest {
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         int expected = 8;
-        int actual = services.maxSales(sales);
+        long actual = services.maxSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -36,7 +36,7 @@ public class StatsServiceTest {
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         int expected = 180;
-        int actual = service.sumSales(sales);
+        long actual = service.sumSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -48,7 +48,7 @@ public class StatsServiceTest {
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         int expected = 15;
-        int actual = service.averageSales(sales);
+        long actual = service.averageSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -60,7 +60,7 @@ public class StatsServiceTest {
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         int expected = 5;
-        int actual = service.moreAverageSales(sales);
+        long actual = service.moreAverageSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -72,7 +72,7 @@ public class StatsServiceTest {
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         int expected = 5;
-        int actual = service.lessAverageSales(sales);
+        long actual = service.lessAverageSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
